execo_g5k api_utils add set_nodes_vlan
The set_nodes_vlan allow to put an arbitrary interface of a list of
hosts in a given vlan. The code makes use of the API to achieve its goal
by sending a POST request to the uri /sites/%site/vlan/%vlan_id and
passing a body composed of the "network_addresses" of the different
(node,interface) to put in the vlan.